Expose run history functionality on all program managers.
