+++ cdap-integration-test/src/main/java/co/cask/cdap/test/remote/RemoteApplicationManager.java
@@ -23,6 +23,7 @@ import co.cask.cdap.client.util.RESTClient;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.ProgramRecord;
 import co.cask.cdap.proto.ProgramType;
+import co.cask.cdap.proto.RunRecord;
 import co.cask.cdap.test.AbstractApplicationManager;
 import co.cask.cdap.test.DataSetManager;
 import co.cask.cdap.test.DefaultMapReduceManager;
@@ -37,6 +38,7 @@ import co.cask.cdap.test.WorkflowManager;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Throwables;
 
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -52,8 +54,8 @@ public class RemoteApplicationManager extends AbstractApplicationManager {
     super(application);
 
     this.clientConfig = clientConfig;
-    this.programClient = new ProgramClient(clientConfig);
-    this.applicationClient = new ApplicationClient(clientConfig);
+    this.programClient = new ProgramClient(clientConfig, restClient);
+    this.applicationClient = new ApplicationClient(clientConfig, restClient);
     this.restClient = restClient;
   }
 
@@ -146,6 +148,15 @@ public class RemoteApplicationManager extends AbstractApplicationManager {
       String status = programClient.getStatus(programId);
       // comparing to hardcoded string is ugly, but this is how appFabricServer works now to support legacy UI
       return "STARTING".equals(status) || "RUNNING".equals(status);
+    } catch (Exception e) {
+      throw Throwables.propagate(e);
+    }
+  }
+
+  @Override
+  public List<RunRecord> getHistory(Id.Program programId) {
+    try {
+      return programClient.getProgramRuns(programId, "ALL", 0, Long.MAX_VALUE, Integer.MAX_VALUE);
     } catch (Exception e) {
       throw Throwables.propagate(e);
     }
