+++ cdap-test/src/main/java/co/cask/cdap/test/AbstractProgramManager.java
@@ -17,8 +17,10 @@
 package co.cask.cdap.test;
 
 import co.cask.cdap.proto.Id;
+import co.cask.cdap.proto.RunRecord;
 import com.google.common.collect.ImmutableMap;
 
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
@@ -101,5 +103,10 @@ public abstract class AbstractProgramManager<T extends ProgramManager> implement
     if (!statusMatched) {
       throw new IllegalStateException("Program state not as expected. Expected " + status);
     }
+  }
+
+  @Override
+  public List<RunRecord> getHistory() {
+    return applicationManager.getHistory(programId);
   }
 }
